+++ docs/changes.md
@@ -1,5 +1,7 @@
 # 最新动态
 
+2026/02/06
+  * 增加 tk_str_ieq_with_len 和 tk_str_find (感谢兆坤提供补丁)
 
 2026/02/06
   * edit_ex 支持按下 enter 键搜索(感谢兆坤提供补丁)